split cv2.cpp: convert
