+++ modules/python/common.cmake
@@ -19,7 +19,16 @@ if(NOT WIN32 AND NOT APPLE AND NOT OPENCV_PYTHON_SKIP_LINKER_EXCLUDE_LIBS)
   set(CMAKE_MODULE_LINKER_FLAGS "${CMAKE_MODULE_LINKER_FLAGS} -Wl,--exclude-libs=ALL")
 endif()
 
-ocv_add_library(${the_module} MODULE ${PYTHON_SOURCE_DIR}/src2/cv2.cpp ${PYTHON_SOURCE_DIR}/src2/cv2_util.cpp ${PYTHON_SOURCE_DIR}/src2/cv2_numpy.cpp ${cv2_generated_hdrs} ${opencv_userdef_hdrs} ${cv2_custom_hdr})
+ocv_add_library(${the_module} MODULE
+  ${PYTHON_SOURCE_DIR}/src2/cv2.cpp
+  ${PYTHON_SOURCE_DIR}/src2/cv2_util.cpp
+  ${PYTHON_SOURCE_DIR}/src2/cv2_numpy.cpp
+  ${PYTHON_SOURCE_DIR}/src2/cv2_convert.cpp
+  ${cv2_generated_hdrs}
+  ${opencv_userdef_hdrs}
+  ${cv2_custom_hdr}
+)
+
 if(TARGET gen_opencv_python_source)
   add_dependencies(${the_module} gen_opencv_python_source)
 endif()
